% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bootStatsMean}
\alias{bootStatsMean}
\title{Bootstrap statistics of sample distribution of the mean}
\usage{
bootStatsMean(x, B = NULL, seed = 42L)
}
\arguments{
\item{x}{Sample of a random variable.}

\item{B}{Bootstrap size for estimating sampling distribution of the mean, default to sample size.}

\item{seed}{Random seed for bootstrapping.}
}
\value{
Data.table of sample mean, standard error, bootstrap range and 90% confidence intervals.
          If B = 100, bootstrap range is ~98% CI.
}
\description{
Bootstrap statistics of sample distribution of the mean
}
