% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{dgpParabola}
\alias{dgpParabola}
\title{Data Generating Process for the Parabola Example}
\usage{
dgpParabola(N)
}
\arguments{
\item{N}{sample size}
}
\description{
y = x^2
x ~ U[0,1]
x' = x + e * (1 + 2 x)
y' = y + e * (1 + 2 y)
}
