% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParameterizedManifolds.R
\name{drLogPNoisyCircle}
\alias{drLogPNoisyCircle}
\title{Partial derivative of the log density in the radial direction.}
\usage{
drLogPNoisyCircle(r, h = 0.2)
}
\description{
Partial derivative of the log density in the radial direction.
}
