% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMapProjection.R
\docType{data}
\name{E}
\alias{E}
\title{Expectations of some random variables.}
\format{
An object of class \code{list} of length 2.
}
\usage{
E
}
\description{
E|z|, absolute value of standard Gaussian.
E(x_2), Rayleigh distribution, or chi with 2 degrees of freedom, x(2).
}
\keyword{internal}
