% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCMS.R
\name{estimateRidge}
\alias{estimateRidge}
\title{Ridge Estimation}
\usage{
estimateRidge(
  X0,
  d,
  sigma,
  minCos = 0.05,
  omega = 1,
  maxStep = NULL,
  maxIter = 100L,
  returnV = FALSE,
  silent = FALSE
)
}
\arguments{
\item{X0}{Initial data.table, N-by-n.}

\item{d}{Dimension of density ridge.}

\item{sigma}{KDE bandwidth for SCMS.}

\item{omega}{Stepsize relaxation factor, default to 1.}

\item{returnV}{Whether to return the eigenvector at ridge; an (n,n,N) array.}
}
\value{
Data.table of estimated density ridge.
}
\description{
Ridge Estimation
}
