% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffusionPropagate.R
\name{gd}
\alias{gd}
\title{Heat diffusion of the probability measure/mass on mesh}
\usage{
gd(M, p0, t, mult = 3, smooth = 1, scale = NULL)
}
\arguments{
\item{M}{Mesh points, an n-by-m' matrix.}

\item{t}{Diffusion time, as "relative time" for the diffusion to reach the whole set.}

\item{mult}{Multiples of the mean square distance within the set, determines the scale parameter.}

\item{smooth}{Multiple of the scale parameter for smoothing density estimate.
#' @return Probability masses on the mesh points.}

\item{p}{Initial probability vector.}
}
\description{
Heat diffusion of the probability measure/mass on mesh
}
