% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParameterizedManifolds.R
\name{logPNoisyCircle}
\alias{logPNoisyCircle}
\title{Log density of the unit circle with Gaussian noise, disregarding the constant term.}
\usage{
logPNoisyCircle(r, h = 0.2)
}
\arguments{
\item{r}{radial coordinate}

\item{h}{bandwidth of Gaussian noise}
}
\description{
Log density of the unit circle with Gaussian noise, disregarding the constant term.
}
