% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MParzen.R
\name{LOOCVMLBandwidth}
\alias{LOOCVMLBandwidth}
\title{Optimal Parzen bandwidth that maximizes log-likelihood on test set.}
\usage{
LOOCVMLBandwidth(X, hlim)
}
\arguments{
\item{X}{Data, data.table.}

\item{hlim}{Range of bandwidth searching.}
}
\value{
Data.table: minimum, optimal bandwidth; objective, average negative log-likelihood (ANLL).
}
\description{
Optimal Parzen bandwidth that maximizes log-likelihood on test set.
}
