% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MParzen.R
\name{LOOCVParzenANLL}
\alias{LOOCVParzenANLL}
\title{Leave-one-out cross validation.}
\usage{
LOOCVParzenANLL(X, sigma)
}
\arguments{
\item{X}{Data matrix, N-by-n.}

\item{sigma}{Marginal standard deviation of an isotropic Gaussian added to each point.
Should be greater than zero.}
}
\value{
Average negative log likelihood by Leave-one-out cross validation,
        given the Parzen model bandwidth.
}
\description{
Leave-one-out cross validation.
}
