% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMapProjection.R
\name{manifoldLearn}
\alias{manifoldLearn}
\title{Manifold learning via diffusion map}
\usage{
manifoldLearn(H0, mult = 3, kappa = 1, delta = 0.001)
}
\arguments{
\item{H0}{The normalized sample matrix, m-by-n.}

\item{mult}{Multiples of the mean square distance within the set, determines the scale parameter.}

\item{kappa}{Analysis scale, as "relative time" for the diffusion to reach the whole set.}

\item{delta}{Convergence criterion}
}
\value{
A list: lambda, eigenvalues of the diffusion map;
                g, diffusion map basis;
                k, minimal truncation order satisfying the convergence criterion.
}
\description{
Manifold learning via diffusion map
}
