% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMapProjection.R
\name{manifoldSample}
\alias{manifoldSample}
\title{Manifold sampling: Gaussian KDE projected on truncated diffusion maps basis.}
\usage{
manifoldSample(H0, g, m, mult = 0L, avgScatter = 0.1, seed = 42L)
}
\arguments{
\item{H0}{The normalized sample matrix, n-by-m.}

\item{g}{Diffusion map basis}

\item{m}{Truncation order}

\item{mult}{Multiple in sample size of the new sample; if 0, project the original sample.}

\item{avgScatter}{Average relative scattering distance from the original sample.}

\item{seed}{Seed for random number generation.}
}
\value{
A list: kde, KDE sample matrix, n-by-m x mult;
                kdem, manifold-reduced sample matrix, n-by-m x mult.
}
\description{
Manifold sampling: Gaussian KDE projected on truncated diffusion maps basis.
}
