% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCMS.R
\name{MeanShift}
\alias{MeanShift}
\title{(KDE-based) Mean shift}
\usage{
MeanShift(X, h, Y0 = NULL, minD = 1e-12, maxIter = 100L, silent = FALSE)
}
\arguments{
\item{X}{Data, an N-by-n matrix.}

\item{h}{Bandwidth for isotropic Gaussian kernel density estimation, a scalar.}

\item{Y0}{Initial points, an M-by-n matrix, defaults to X.}

\item{minD}{Convergence criterion, length of mean shift vector.}

\item{maxIter}{Maximum number of iterations.}
}
\value{
A list: `Y`, the final points; `updatedPoints`, the number of updated points per update.
}
\description{
(KDE-based) Mean shift
}
\references{
@Comaniciu2002
}
