% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MParzen.R
\name{MLBandwidth}
\alias{MLBandwidth}
\title{Optimal Parzen bandwidth that maximizes log-likelihood on test set.}
\usage{
MLBandwidth(XTrain, XTest, hlim)
}
\arguments{
\item{XTrain}{Training data.table.}

\item{XTest}{Test data.table.}

\item{hlim}{Range of bandwidth searching.}
}
\value{
Data.table: minimum, optimal bandwidth; objective, average negative log-likelihood (ANLL).
}
\description{
Optimal Parzen bandwidth that maximizes log-likelihood on test set.
}
