% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mode1d}
\alias{mode1d}
\title{Mode estimation in 1d}
\usage{
mode1d(x, minD, silent = FALSE)
}
\arguments{
\item{x}{Data in 1d, a vector.}

\item{minD}{Convergence criterion for mean shift.}
}
\value{
Estimated mode.
}
\description{
Assumes a unimodal distribution.
}
