% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MParzen.R
\name{MParzenANLL}
\alias{MParzenANLL}
\title{Average `MParzenTest()` over all test points}
\usage{
MParzenANLL(Y, X, d, k, sigma)
}
\arguments{
\item{Y}{Matrix of test data, n columns.}

\item{X}{Matrix of training data, n columns.}

\item{d}{Dimension of manifold.}

\item{k}{Number of nearest neighbors to use in local covariance analysis.}

\item{sigma}{Maximum noise level added to the normal space of each kernel,
i.e. marginal standard deviation of an isotropic Gaussian. Should be greater than zero.}
}
\value{
Average negative log likelihood of the test data given the MParzen model
        on the training data and the hyperparameters.
}
\description{
Useful for tuning hyperparameters: d, k, sigma.
}
