% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MParzen.R
\name{MParzenRadiusTrain}
\alias{MParzenRadiusTrain}
\title{Train MParzen model with fixed-radius nearest neighbors}
\usage{
MParzenRadiusTrain(X, d, r)
}
\description{
Train MParzen model with fixed-radius nearest neighbors
}
