% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MParzen.R
\name{MParzenSampleNoNoise}
\alias{MParzenSampleNoNoise}
\title{Sample from MParzen model on tangent spaces}
\usage{
MParzenSampleNoNoise(N, X, model)
}
\arguments{
\item{N}{Number of samples.}

\item{X}{matrix of training data.}

\item{model}{Top d-th order local covariance structure; output of `MParzenTrain()`.}
}
\value{
A sample of size N from the Manifold Parzen model with no noise.
}
\description{
65% time than `MParzenSample(sigma = 0)`.
}
