% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MParzen.R
\name{MParzenTrain}
\alias{MParzenTrain}
\title{Training MParzen model}
\usage{
MParzenTrain(X, d, k)
}
\arguments{
\item{X}{Matrix of training data, l-by-n.}

\item{d}{Dimension of manifold.}

\item{k}{Number of nearest neighbors to use in local covariance analysis.}
}
\value{
A list of a (d, l) matrix `sdev` and an (n, d, l) array `tangent`.
        Top d standard deviations and directions of local covariance structure at each point.
}
\description{
Modified to inlcude the point under consideration and center the local data matrix,
instead of center at the point.
}
\references{
@Vincent2002
}
