% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalBundleBootstrap.R
\name{NBB}
\alias{NBB}
\title{Normal-bundle bootstrap.}
\usage{
NBB(X0, sigma = NULL, d = 1L, knn = NULL)
}
\arguments{
\item{X0}{Initial data/sample, a data.table: x, y.}

\item{sigma}{KDE bandwidth for SCMS dynamical system.}

\item{d}{Dimension of density ridge.}

\item{knn}{Number of nearest neighbors to use, default to all.}
}
\value{
New data: I, index of ridge point; k, order of projection vector by distance on ridge.
}
\description{
A direct implementation of NBB without using smooth frame.
}
\references{
Algorithm 3.1 in [@ZhangRD2021nbb].
}
\seealso{
\code{\link{scms}} for ridge estimation;
\code{\link{SmoothFrame}} for smooth frame construction.
}
