% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCMS.R
\name{outlier}
\alias{outlier}
\title{Detect Outers}
\usage{
outlier(X, sds = 6)
}
\arguments{
\item{X}{Data, an N-by-n matrix.}

\item{sds}{Multiple of standard deviations away from the median distance to point cloud
for a data point to be outlier.}
}
\value{
Logical vector indicating outliers in data for a smoother KDE.
}
\description{
Detect Outers
}
\references{
@Genovese2014
}
