% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MParzen.R
\name{ParzenANLL}
\alias{ParzenANLL}
\title{Average `ParzenTest()` over all test points}
\usage{
ParzenANLL(Y, X, sigma)
}
\arguments{
\item{Y}{Matrix of test data, n columns.}

\item{X}{Matrix of training data, n columns.}

\item{sigma}{Marginal standard deviation of an isotropic Gaussian added to each point.
Should be greater than zero.}
}
\value{
Average negative log likelihood of the test data given the Parzen model
        on the training data and the bandwidth.
}
\description{
Useful for tuning hyperparameters: d, k, sigma.
}
