% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MParzen.R
\name{ParzenTest}
\alias{ParzenTest}
\title{Parzen model}
\usage{
ParzenTest(y, X, sigma)
}
\arguments{
\item{y}{density evaluation point.}

\item{X}{matrix of training data.}

\item{sigma}{Marginal standard deviation of an isotropic Gaussian added to each point.
Should be greater than zero.}
}
\value{
Parzen estimator at the evaluation point.
}
\description{
Parzen model
}
