% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plmR.R
\name{plqrss.fit}
\alias{plqrss.fit}
\title{Partial Linear Quantile Smoothing Spline for a High-dimensional Heterogeneous Data}
\usage{
plqrss.fit(
  y,
  x,
  z,
  tau,
  lambda1,
  lambda2,
  iter.max = 1000,
  w.iter.max = 100,
  tol1 = 0.001,
  tol2 = 1e-04,
  init.g = NULL,
  ...
)
}
\arguments{
\item{y}{a vector of a response variable for the model.}

\item{x}{a matrix of covariates related to the linear part.}

\item{z}{a vector of covariate related to the linear part.}

\item{tau}{a real number for a quantile level ranged between 0 and 1.}

\item{lambda1}{a real number which is the tuning parameter for LASSO}

\item{lambda2}{a real number which is the tuning parameter for trend filering}

\item{iter.max}{a maximum iteration numer for the blockwise coordinate descent (BCD).}

\item{w.iter.max}{a maximum iteration numer for the MM algorithm for quantile regression.}

\item{tol1}{a tolerance error when the BCD loop break.}

\item{tol2}{a tolerance error when the MM loop break.}

\item{init.g}{(optional) a vector used for initialization of nonparametric term.}

\item{...}{used for controlling the others.}
}
\value{
\code{plqrss.fit} returns an object of class \code{"list"}, which is a list containing
following components:

\item{beta}{fitted coefficients for the linear term.}
\item{g}{fitted values for nonparametric terms.}
\item{model.g}{fitted nonparmetric object.}
}
\description{
The \code{plqrss.fit} performs partial linear quantile trend filtering model on the high-dimensional data by combining Quantile LASSO and smoothing spline.
}
\author{
Sang Kyu Lee
}
