% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plmR.R
\name{plss.fit}
\alias{plss.fit}
\title{Partial Linear Smoothing Spline for a High-dimensional Heterogeneous Data}
\usage{
plss.fit(y, x, z, lambda1, lambda2, iter.max = 1000, tol1 = 0.001, ...)
}
\arguments{
\item{y}{a vector of a response variable for the model.}

\item{x}{a matrix of covariates related to the linear part.}

\item{z}{a vector of covariate related to the linear part.}

\item{lambda1}{a real number which is the tuning parameter for LASSO}

\item{lambda2}{a real number which is the tuning parameter for trend filering}

\item{iter.max}{a maximum iteration numer for the blockwise coordinate descent (BCD).}

\item{tol1}{a tolerance error when the BCD loop break.}

\item{...}{used for controlling the others.}
}
\value{
\code{plss.fit} returns an object of class \code{"list"}, which is a list containing
following components:

\item{beta}{fitted coefficients for the linear term.}
\item{g}{fitted values for nonparametric terms.}
\item{model.g}{fitted nonparmetric object.}
}
\description{
The \code{plss.fit} performs partial linear quantile trend filtering model on the high-dimensional data by combining Quantile LASSO and smoothing spline.
}
\author{
Sang Kyu Lee
}
