% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffusionPropagate.R
\name{pmesh}
\alias{pmesh}
\title{Estimate probability measure on mesh by fine Gaussian kernels}
\usage{
pmesh(M, X, mult = 1/2)
}
\arguments{
\item{M}{Mesh points, an n-by-m' matrix.}

\item{X}{Sample data, an n-by-m matrix.}

\item{mult}{Multiples of the mean square distance within the set, determines the scale parameter.}
}
\value{
A list: p, probability mass on mesh; f, probability density on mesh.
}
\description{
Estimate probability measure on mesh by fine Gaussian kernels
}
