% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMapProjection.R
\name{preprocess}
\alias{preprocess}
\title{Preprocessing}
\usage{
preprocess(X0, scale = FALSE)
}
\arguments{
\item{X0}{The original sample matrix, m-by-n; can be a data.frame.}

\item{scale}{A logical for whether the variables should be scaled.}
}
\value{
A list: pc, affine transformation to normalized principal component (PC) coordinates;
                H0, normalized sample matrix (zero-mean, unit-variance), m-by-n;
}
\description{
Preprocessing
}
\details{
All PC coordinates are currently assumed to be non-degenerate.
         Further generalization is needed.
}
