% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{radial}
\alias{radial}
\title{Radial coordinate of the parabola example.}
\usage{
radial(x, y)
}
\value{
radial angle of input points, between 0 and pi, mostly around pi/2.
}
\description{
The radial partitions the data set rather evenly.
}
