% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParameterizedManifolds.R
\name{rCircularChi}
\alias{rCircularChi}
\title{Sampling on a circular chi distribution in a flow model.}
\usage{
rCircularChi(N, df)
}
\arguments{
\item{N}{Sample size.}

\item{df}{Degree of freedom of the chi distribution; df >= 2, integer.}
}
\value{
Data.table: theta, azimuth angle; r, radial coordinate; x, y, Cartesian coordinates.
}
\description{
Circular chi: uniform distribution in azimuth angle,
mode-normalized chi distribution in the radial coordinate.
Density ridge is r=1.
}
