% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParameterizedManifolds.R
\name{rCircularGauss}
\alias{rCircularGauss}
\title{Sampling on a circular Gaussian distribution.}
\usage{
rCircularGauss(N, sigma)
}
\arguments{
\item{N}{Sample size.}

\item{sigma}{Standard deviation of Gaussian noise in the radial direction.}
}
\value{
Data.table: theta, azimuth angle; r, radial coordinate; x, y, Cartesian coordinates.
}
\description{
Sampling on a circular Gaussian distribution.
}
