% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMapProjection.R
\name{revpredict}
\alias{revpredict}
\title{Reverse prediction: principal components to original coordinates}
\usage{
revpredict(H, pc)
}
\arguments{
\item{H}{Principal component coordinates, m-by-n.}

\item{pc}{Affine transformation to principal component coordinates.}

\item{scale}{Whether restore original scale.}
}
\value{
Original coordinates, m-by-n.
}
\description{
Reverse prediction: principal components to original coordinates
}
