% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParameterizedManifolds.R
\name{rhelix1}
\alias{rhelix1}
\title{Helix 1: cylindrical coordinates}
\usage{
rhelix1(N, invCDF, a = 2, b = 1, n = 7)
}
\arguments{
\item{N}{Sample size.}

\item{invCDF}{Inverse CDF, for probability integral transformation.}

\item{a}{Major radius.}

\item{b}{Minor radius.}

\item{n}{Number of periods.}
}
\description{
Helix 1: cylindrical coordinates
}
