% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParameterizedManifolds.R
\name{rsphere}
\alias{rsphere}
\title{Sampling from a distribution on the unit sphere}
\usage{
rsphere(N, f)
}
\arguments{
\item{N}{Sample size.}

\item{f}{Probability density as a function of spherical coordinates (theta, phi).}
}
\description{
Sampling from a distribution on the unit sphere
}
