% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParameterizedManifolds.R
\name{rtorus}
\alias{rtorus}
\title{Sampling from a distribution on a torus}
\usage{
rtorus(N, f, a, b)
}
\arguments{
\item{N}{Sample size.}

\item{f}{Probability density as a function of spherical coordinates (theta, phi).}

\item{a}{Radius of big circle.}

\item{b}{Radius of small circle.}
}
\description{
Sampling from a distribution on a torus
}
