% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParameterizedManifolds.R
\name{runifcircle}
\alias{runifcircle}
\title{Sampling uniformly on the unit circle.}
\usage{
runifcircle(N)
}
\arguments{
\item{N}{Sample size.}
}
\value{
Data.table: theta, azimuth angle; x, y, Cartesian coordinates.
}
\description{
Sampling uniformly on the unit circle.
}
