% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParameterizedManifolds.R
\name{runifsphere}
\alias{runifsphere}
\title{Sampling uniformly on the unit sphere, as a flow model.}
\usage{
runifsphere(N)
}
\arguments{
\item{N}{Sample size.}
}
\value{
Data.table: theta, azimuth angle; phi, polar angle.
}
\description{
Sampling uniformly on the unit sphere, as a flow model.
}
