% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCMS.R
\name{scms}
\alias{scms}
\title{KDE-based Subspace Constrained Mean Shift}
\usage{
scms(
  Y0,
  X,
  d,
  h,
  log = TRUE,
  r = NULL,
  omega = 1,
  maxStep = NULL,
  minCos = 0.01,
  maxIter = 100L,
  returnV = FALSE,
  silent = FALSE
)
}
\arguments{
\item{Y0}{Initial points, an n-by-M matrix.}

\item{X}{Data, an n-by-N matrix.}

\item{d}{Dimension of the density ridge, an integer.}

\item{h}{Bandwidth for isotropic Gaussian kernel density estimation, a scalar.}

\item{log}{Whether to use the log Gaussian KDE. Default to TRUE.}

\item{r}{Radius of nearest neighbors to use in evaluating density and derivatives.
Default to use all data.}

\item{omega}{Stepsize relaxation factor.}

\item{maxStep}{Maximum stepsize. You probably don't need to use both `omega` and `maxStep`.}

\item{minCos}{Convergence criterion, cosine of the angle between gradient and
its component in the "local normal space".}

\item{maxIter}{Maximum number of iterations.}

\item{returnV}{Whether to return the eigenvector at Y; an (n,n,M) array.}
}
\value{
A list: `Y`, the final points; `cosNormal`, convergence score;
                `lambdaNormal`, largest Hessian eigenvalue in the normal space;
                `eigengap`, d-th eigengap of the Hessian;
                `updatedPoints`, the number of updated points per update;
}
\description{
Compuational complexity per step: O(M * N * n^3), where M is for every query point,
N for every data point, n^3 for eigen-decomposition.
If partial/truncated eigen-decomposition is used, the n^3 term would become d*n^2
(e.g. the Lanczos algorithm, which further reduces by the sparsity of the matrix).
}
\details{
In the return, `lambdaNormal` and `eigengap` are both useful for checking
if the final point is a ridge point.
}
\references{
@Ozertem2011
Update a point by index.
Return new coordinates and cosine between gradient and local normal space.
}
