% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCMS.R
\name{scmsMParzen}
\alias{scmsMParzen}
\title{MParzen KDE-based Subspace Constrained Mean Shift}
\usage{
scmsMParzen(Y0, X, model, sigma, h, r = NULL, minCos = 0.01, maxIter = 100L)
}
\arguments{
\item{Y0}{Initial points, an n-by-M matrix.}

\item{X}{Data, an n-by-N matrix.}

\item{model}{Top d-th order local covariance structure; output of `MParzenTrain()`.}

\item{sigma}{Maximum noise level added to the normal space of each kernel,
i.e. marginal standard deviation of an isotropic Gaussian. Can be zero.}

\item{h}{Bandwidth for isotropic Gaussian kernel density estimation, a scalar.
Determines step size.}

\item{r}{Radius of nearest neighbors to use in evaluating density and derivatives.
Default to use all data.}

\item{minCos}{Convergence criterion, cosine of the angle between gradient and
its component in the "local normal space".}

\item{maxIter}{Maximum number of iterations.}
}
\value{
A list: `Y`, the final points; `cosNormal`, convergence score;
                `updatedPoints`, the number of updated points per update.
}
\description{
Using the log Manifold Parzen windows (MParzen) Gaussian KDE.
}
