% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCMS.R
\name{silverman}
\alias{silverman}
\title{Silverman's rule of thumb bandwidths}
\usage{
silverman(N, n, sd)
}
\arguments{
\item{N}{Sample size.}

\item{n}{Dimension of observation.}

\item{sd}{Data standard deviations.}
}
\description{
The optimal bandwidths for Gaussian KDE on a Gaussian random vector
in terms of the mean integrated squared error (MISE).
}
\references{
@Silverman1986
}
