% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmoothFrame.R
\name{SmoothFrame}
\alias{SmoothFrame}
\title{Construct a smooth frame from an initial orthonormal frame}
\usage{
SmoothFrame(V, Mridge, j = NULL)
}
\arguments{
\item{V}{Initial orthonormal frame, an (n, d, N) array for a rank-d bundle.}

\item{Mridge}{Ridge points, an N-by-n matrix.}

\item{j}{Index of the reference frame.}
}
\description{
A general function that constructs a smooth frame of a vector bundle over a smooth manifold.
It can be applied to the normal or tangent bundle of an estimated density ridge.
}
\references{
See Supplementary Materials of [@ZhangRD2021nbb].
For the `align` sub-procedure, see [@Rheinboldt1988] cited therein.
}
